Rename proxy.settings to network.settings
Also clean up the code in NetworkingConfigurator a bit.
