+++ components/scripts/gradle/network.settings
