+++ components/scripts/lib/build_scan.sh
@@ -166,7 +166,7 @@ fetch_and_read_build_scan_data() {
     args+=(-m "${mapping_file}")
   fi
 
-  args+=("--proxy-settings-file" "${SCRIPT_DIR}/proxy.settings")
+  args+=("--network-settings-file" "${SCRIPT_DIR}/network.settings")
 
   if [[ "$1" == "build_cache_metrics_only" ]]; then
     build_cache_metrics_only="true"
