+++ components/scripts/maven/network.settings
